+++ tech-motisha/src/App.css
@@ -1,30 +1,30 @@
-/* styles for all devices */
+/* styles for all devices
 @import url('https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css');
 body {
     font-size: 16px;
   }
   
   /* styles for devices with a max width of 768px */
-  @media (max-width: 768px) {
+  /* @media (max-width: 768px) {
     body {
       font-size: 14px;
     }
-  }
+  } */
   
   /* styles for devices with a min width of 1024px */
-  @media (min-width: 1024px) {
+  /* @media (min-width: 1024px) {
     body {
       font-size: 18px;
     }
-  }
+  } */
 
 :root{
    --black: rgba(0,0,0,1);
-   --bunting: rgba(14,33,69,1);
+   --bunting: #0e2145;
    --bunting-2:rgba(16,31,60,1);
-   --alto: rgba(217,217,217,1);
+   --alto: #d9d9d9;
    --oslo-gray: rgba(136,137,149,1);
-   --international-orange: rgba(250,81,15,1);
+   --international-orange: #fa510f;
    --white: rgba(255,255,255,1);
 
    --font-size-s: 1.25rem; /*20px*/
