+++ tech-motisha/tailwind.config.js
@@ -2,6 +2,15 @@
 module.exports = {
   content: ["./src/**/*.{html,js}"],
   theme: {
+    colors: {
+      transparent: 'transparent',
+      current: 'currentColor',
+      'orange': '#fa510f',
+      'bunting-2':'rgba(16,31,60,1)',
+      'bunting': '#0e2145',
+      'alto': '#d9d9d9'
+
+    },
     extend: {},
   },
   plugins: [],
